code shifted chat to input bar
